Refactor miniWeather
* Separate out all the loops that can be parallelized
  into their own functions

* miniWeather_mdspan.cpp is the "generic executable"
